This commit includes the first unit test added to the solution, using xUnit. The approach demonstrates some fundamentals about working with tests in EF Core and AspNetCore, which will be helpful to observe when doing integration tests as well. Moq is being used in the unit test to mock out the data context, providing it test data that the controller's logic can call against.
A class called TestAsyncQueryProvider has been added as well. This is a very
important class that serves as a test double of the provider that will process
asynchronous queries during the unit tests. It is decorated with some comments
and source URLs for further details. For API controllers, we may need to
create a similar object, though I'm hoping this will work for those as well.

For now, just a single test is listed, so that we can look at it together and
discuss. There's a single line in the Assert portion of the test that is
commented out, as it represents a change from EF6 and its syntax is troubling
me. I would like to be able to verify that the method
Labels.FirstOrDefaultAsync was called on the mocked data context. Obviously,
it is getting called, but a change in syntax from the old FindAsync(int) to
this new one, which expects an expression makes it difficult to understand how
to tell Moq to verify it. Any help there would be great!