IEfCoreMsSqlRepository and a generic implementation EfCoreMsSqlRepository<T> have been added to the Data project, to serve as a generic repository for interacting with EF Core's MS SQL databases.
A unit test suite has also been added to test cover the generic repository.
It is currently located in the same unit tests project as the tests for
controllers, as that project has all the necessary dependencies for testing
already. We probably should now create a second unit test project, specific to
the Tuneage.Data project, with all the same dependencies, and move this test
there.