Integration tests added, with in-memory database usage, and data seeding from our TestDataGraph. This introduces new objects TestStartup (inherits from Startup but overrides some methods to configure an in-memory database specifically for testing) and DataSeeder (used by the TestStartup's Configure method to seed data to the in-memory database).
These tests only cover the MVC routes, so API tests need to be added next.
Also, the POST tests are failing, returing Bad Request. There must be more
setup needed to call those methods correctly, so the tests that call PostAsync
on the client are currently ignored, pending a solution to that.